separate asset, entry, and richtext functions
